Flash Message

+++ DiarioEscolar/Models/NotaFalta.cs
@@ -14,19 +14,19 @@ namespace DiarioEscolar.Models
 
         public decimal Nota1{ get; set; }
 
-        public decimal Falta1 { get; set; }
+        public int Falta1 { get; set; }
 
         public decimal Nota2 { get; set; }
 
-        public decimal Falta2 { get; set; }
+        public int Falta2 { get; set; }
 
         public decimal Nota3 { get; set; }
 
-        public decimal Falta3 { get; set; }
+        public int Falta3 { get; set; }
 
         public decimal Nota4 { get; set; }
 
-        public decimal Falta4 { get; set; }
+        public int Falta4 { get; set; }
         
         public decimal Recuperacao { get; set; }
